+++ Changelog.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change export method in utils.js. Completed on 2024-09-11
 - Change the grab the remaining string to instead the older method in establish/resource method at deskelectron/src/desktop.js. Completed on 2024-09-18
 - Apply libsql node module instead the better-sqlite3 which incompatible with bun.js. This affect webnodejs and slqmanager engine. Completed on 2024-09-18
+- Re-structure the module import method. Affected backend code includes guimaker,smfetch,deskelectronjs,sqlmanager and webnodejs. Completed on 2024-09-24
 
 ### Deprecated